tested LinearRegressionMulticlass

+++ ml_core/classification/logistic_functions/__init__.py
@@ -3,4 +3,4 @@ import numpy as np
 
 def sigmoid(x: np.ndarray) -> np.ndarray:
 
-	return 1 / (1 + np.exp(-x))
+	return 1 / (1 + np.exp(-x.astype(float)))
